+++ Source/Config.cs
@@ -21,7 +21,6 @@ public class Config(ConfigFile file)
     public ConfigEntry<bool> EnableDynamicResolution { get; } = file.Bind("Performance", "EnableDynamicResolution", false, "Whether or not dynamic resolution should be enabled. Required for most of these settings to have an effect.");
     public ConfigEntry<DynamicResUpscaleFilter> DynamicResolutionUpscaleFilter { get; } = file.Bind("Performance", "DynamicResolutionUpscaleFilter", DynamicResUpscaleFilter.EdgeAdaptiveScalingUpres, new ConfigDescription("The filter/algorithm that will be used to perform dynamic resolution upscaling. Defaulted to FSR (Edge Adaptive Scaling).", new AcceptableValueEnum<DynamicResUpscaleFilter>()));
     public ConfigEntry<float> DynamicResolutionPercentage { get; } = file.Bind("Performance", "DynamicResolutionPercentage", 80f, new ConfigDescription("The percentage of resolution to scale the game down to. The lower the value, the harder the upscale filter has to work which will result in quality loss.", new AcceptableValueRange<float>(0, 100)));
-    public ConfigEntry<bool> EnableDLSS { get; } = file.Bind("Performance", "EnableDLSS", false, "[DEPRECATED] DLSS support will be removed in a future release!");
     public ConfigEntry<float> CameraResolution { get; } = file.Bind("Performance", "CameraResolution", 0.75f, new ConfigDescription("This setting configures the resolution scale of the game, lower values are more performant, but will make the game look worse. From around 0.8 the difference is negligible (on a Quest 2, with dynamic resolution disabled).", new AcceptableValueRange<float>(0.05f, 1f)));
     public ConfigEntry<bool> DisableVolumetrics { get; } = file.Bind("Performance", "DisableVolumetrics", false, "Disables volumetrics in the game, which significantly improves performance, but removes all fog and may be considered cheating.");
 
@@ -59,12 +58,19 @@ public class Config(ConfigFile file)
     public ConfigEntry<bool> DisableCompanyBellInteraction { get; } = file.Bind("Interaction", "DisableCompanyBellInteraction", false, "Disables needing to physically press the bell at the company desk.");
     public ConfigEntry<bool> DisableBreakerBoxInteraction { get; } = file.Bind("Interaction", "DisableBreakerBoxInteraction", false, "Disabled needing to physically open the breaker box and flip the switches with your finger.");
     public ConfigEntry<bool> DisableDoorInteraction { get; } = file.Bind("Interaction", "DisableDoorInteraction", false, "Disable needing to physically open and close doors by interacting with the door handles. Will also disable the need to use keys and lockpickers physically on the door handle.");
-
     public ConfigEntry<bool> DisableHangarLeverInteraction { get; } = file.Bind("Interaction", "DisableHangarLeverInteraction", false, "Disable needing to physically pull the lever for the big doors on Artiface");
-    
     public ConfigEntry<bool> DisableMuffleInteraction { get; } = file.Bind("Interaction", "DisableMuffleInteraction", false, "Disables the self-muffling feature, which makes it so that holding your hand in front of your mouth will no longer make you inaudible to enemies.");
     public ConfigEntry<bool> DisableFaceInteractions { get; } = file.Bind("Interaction", "DisableFaceInteractions", false, "Disables the functionality to hold certain items up to your face to use them.");
 
+    // Car interaction configuration
+
+    public ConfigEntry<bool> DisableCarSteeringWheelInteraction { get; } = file.Bind("Car", "DisableCarSteeringWheelInteraction", false, "Disables the need to physically steer the Company Cruiser");
+    public ConfigEntry<bool> DisableCarButtonInteractions { get; } = file.Bind("Car", "DisableCarButtonInteractions", false, "Disables the need to physically press the generic buttons in the Company Cruiser (radio, windshield wipers, etc)");
+    public ConfigEntry<bool> DisableCarHonkInteraction { get; } = file.Bind("Car", "DisableCarHonkInteraction", false, "Disables the need to physically press the car honk in the Company Cruiser");
+    public ConfigEntry<bool> DisableCarEjectInteraction { get; } = file.Bind("Car", "DisableCarEjectInteraction", false, "Disables the need to physically press the eject button in the Company Cruiser");
+    public ConfigEntry<bool> DisableCarGearStickInteractions { get; } = file.Bind("Car", "DisableCarGearStickInteractions", false, "Disables the need to physically shift the gears in the Company Cruiser");
+    public ConfigEntry<bool> DisableCarIgnitionInteractions { get; } = file.Bind("Car", "DisableCarIgnitionInteractions", false, "Disables the need to physically start/stop the car ignition in the Company Cruiser");
+    
     // Internal configuration
 
     public ConfigEntry<string> ControllerBindingsOverride { get; } = file.Bind("Internal", "ControllerBindingsOverride", "", "FOR INTERNAL USE ONLY, DO NOT EDIT");